anomaly_tesis_library/requirements.txt
awswrangler==3.2.1
category-encoders==2.6.1
numba==0.57.1
numpy==1.24.4
pandas==2.0.3
pyarrow==12.0.1
requests==2.31.0
s3fs==0.4.2
scikit-learn==1.3.0
seaborn==0.12.2
statsmodels==0.14.0
Unidecode==1.3.6
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==40.0.2
pyod==1.1.0
tensorflow==2.11.1
openpyxl==3.1.2
python-dotenv==1.0.0
tqdm